refactor(codegen): split codegen.service.ts by target-language family
The 2,010-line service held ~50 top-level generator functions for every
target framework plus the CodeGenService class. Generators now live in
src/services/codegen/ grouped by language family (typescript.ts,
java.ts, python.ts, go.ts) with spec-parser.ts, targets.ts and
shared.ts for the cross-family pieces; codegen.service.ts keeps only
the CodeGenService class (~500 lines) with an unchanged public surface.

Pure move-refactor: functions moved verbatim, only export/import
wiring added. 89/89 codegen tests pass, typecheck clean.
